% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_pression_matrix.R
\name{compute_fpr}
\alias{compute_fpr}
\title{Compute False Positive Rate (FPR)}
\usage{
compute_fpr(true_precision, estimated_precision)
}
\arguments{
\item{true_precision}{The ground truth precision matrix.}

\item{estimated_precision}{The estimated precision matrix.}
}
\value{
A numeric value representing the FPR. Returns \code{0} if there
are no true negatives to evaluate.
}
\description{
Calculates the False Positive Rate for an estimated precision matrix
compared to the ground truth structure.
}
\examples{
true_mtx <- diag(1, 3)
est_mtx <- matrix(c(1, 0.5, 0, 0.5, 1, 0, 0, 0, 1), 3, 3)
compute_fpr(true_mtx, est_mtx)
}
