% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{E_tau_given_X}
\alias{E_tau_given_X}
\title{Compute E\link{tau | X = x} for the MGIG Distribution}
\usage{
E_tau_given_X(psi, chi, d, qx)
}
\arguments{
\item{psi}{Numeric scalar. Must be strictly positive.}

\item{chi}{Numeric scalar.}

\item{d}{Numeric scalar. The dimension parameter. Must be \code{>= 2}.}

\item{qx}{Numeric vector or scalar. Evaluated quadratic form values.}
}
\value{
A numeric vector of expected values.
}
\description{
Compute E\link{tau | X = x} for the MGIG Distribution
}
