% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{gamma_glasso}
\alias{gamma_glasso}
\title{Gamma-Lasso Precision Matrix Estimation}
\usage{
gamma_glasso(X, lambda, gamma, max_iter = 100, tol = 0.001, appr = FALSE)
}
\arguments{
\item{X}{Numeric matrix. The \code{n x p} data matrix.}

\item{lambda}{Numeric scalar. L1 penalty parameter for sparsity.}

\item{gamma}{Numeric scalar. Robustness parameter. If \code{0}, it is equivalent to the standard Graphical Lasso.}

\item{max_iter}{Integer. Maximum number of iterations. Default is \code{100}.}

\item{tol}{Numeric scalar. Tolerance for convergence. Default is \code{1e-3}.}

\item{appr}{Logical. Whether to use the approximate formulation in \code{glasso}. Default is \code{FALSE}.}
}
\value{
A list containing the estimated precision matrix (\code{Theta}), the robust covariance matrix (\code{S_robust}), and the robust observation \code{weights}.
}
\description{
Estimates a robust precision matrix using the Gamma-divergence framework.
}
