% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{get_offdiag}
\alias{get_offdiag}
\title{Extract Lower Triangular Off-Diagonal Elements}
\usage{
get_offdiag(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
A vector containing the lower triangular elements.
}
\description{
Extract Lower Triangular Off-Diagonal Elements
}
