% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{gslope_admm}
\alias{gslope_admm}
\title{ADMM Solver for the Gslope Problem}
\usage{
gslope_admm(S, lambda_seq, rho = 1, max_iter = 100, tol = 1e-04)
}
\arguments{
\item{S}{Numeric matrix. The sample covariance matrix.}

\item{lambda_seq}{Numeric vector. Sequence of decreasing penalty values.}

\item{rho}{Numeric scalar. ADMM step size parameter. Default is \code{1.0}.}

\item{max_iter}{Integer. Maximum number of ADMM iterations. Default is \code{100}.}

\item{tol}{Numeric scalar. Tolerance for convergence. Default is \code{1e-4}.}
}
\value{
A numeric matrix representing the estimated precision matrix.
}
\description{
Solves the weighted graphical model problem with a sorted L1 penalty.
}
