% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{Roc_result}
\alias{Roc_result}
\title{Compute Interpolated ROC Curve Results}
\usage{
Roc_result(
  result,
  true_result,
  tol = 0.001,
  fpr_seq = seq(0, 1, length.out = 100)
)
}
\arguments{
\item{result}{Numeric matrix. The estimated precision matrix.}

\item{true_result}{Numeric matrix. The ground truth precision matrix.}

\item{tol}{Numeric scalar. Tolerance threshold for identifying non-zero true edges. Default is \code{1e-3}.}

\item{fpr_seq}{Numeric vector. Sequence of FPR values at which to interpolate TPR.}
}
\value{
A numeric vector of interpolated TPR values.
}
\description{
Compute Interpolated ROC Curve Results
}
