\name{sortReducedResultList}
\alias{sortReducedResultList}
\title{Sorts a list of \code{\link{cv.1l.emlasso.reduced}} by lambda}
\usage{
  sortReducedResultList(reducedResultList)
}
\arguments{
  \item{reducedResultList}{list of
  \code{\link{cv.1l.emlasso.reduced}}}
}
\value{
  similar list, but now sorted by lambda, and of class
  "reducedResultList"
}
\description{
  Sorts a list of \code{\link{cv.1l.emlasso.reduced}} by
  lambda
}
\note{
  used internally by
  \code{\link{getSortedReducedResultList}}
}
\author{
  Nick Sabbe \email{nick.sabbe@ugent.be}
}
\keyword{list}
\keyword{reduce}
\keyword{sort}

