% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{tslope}
\alias{tslope}
\title{Robust TSLOPE Estimator}
\usage{
tslope(X, nu = 3, max_iter_em = 20, tol_em = 0.001)
}
\arguments{
\item{X}{Numeric matrix. The \code{n x p} data matrix.}

\item{nu}{Numeric scalar. Degrees of freedom for the Student-t distribution. Default is \code{3}.}

\item{max_iter_em}{Integer. Maximum number of EM iterations. Default is \code{20}.}

\item{tol_em}{Numeric scalar. Tolerance for EM convergence. Default is \code{1e-3}.}
}
\value{
A list containing the estimated precision matrix (\code{Theta}) and the observation \code{Weights}.
}
\description{
Fits the Graphical SLOPE model using a Student-t EM framework for robust estimation.
}
