\docType{package}
\name{EMLasso-package}
\alias{EMLasso-package}
\title{EM (logistic) LASSO}
\description{
  Fit (logistic) LASSO when there is missing data in the
  predictors
}
\details{
  \tabular{ll}{ Package: \tab EMLasso\cr Type: \tab
  Package\cr Version: \tab 2.0.0\cr Date: \tab
  2012-03-16\cr License: \tab GPL (>= 2)\cr LazyLoad: \tab
  yes\cr Depends: Matrix, glmnet, snowfall, addendum,
  NumDfr, GLoMo\cr}
}
\author{
  Nick Sabbe \email{nick.sabbe@ugent.be}

  Maintainer: Nick Sabbe \email{nick.sabbe@ugent.be}
}
\references{
  [PENDING]
}
\keyword{package}

